Multiprocessing digest, WIP: fixing typo in beast.py, passing meta fields to the digest classes, renaming SingleThreadedDigestor to SingleProcessDigestor to reflect the nature of the class, renaming Meta namedtuple to meta_cls for now, factoring out common methods and routines into the digest/abstract.py and AbstractDigestor class, switching to id2id mapping for the context_entities, switching from EntityProxy to deflate/inflate routines to overcome the pickling issue of the original FTM when doing multiprocessing, adding an implementation for MultiProcessDigestor (via multiprocessing and custom initializer/initargs to pass some arguments and initialize FTM/meta_cls for each worker process according to the mapping), renaming parent_context_entities to parent_context_entities_map, adding some helper functions for iterables, adding new mapping for the multiprocessing digest, removing unused test mapping, adding test for MultiProcessDigestor digest
