+++ beast.py
@@ -4,10 +4,10 @@ from tqdm import tqdm
 import sys
 
 if __name__ == "__main__":
-    mapping = SourceMapping(Path("thebeast/tests/sample/mappings/ukrainian_mps.yaml"))
+    mapping = SourceMapping(Path("thebeast/tests/sample/mappings/ukrainian_mps_multithreaded.yaml"))
 
     mapping.dumper.write_entities(
-        tqdm(mapping.digestor.extract(tqdm(mapping.ingestor, desc="Records in")), desc="Entities outs")
+        tqdm(mapping.digestor.extract(tqdm(mapping.ingestor, desc="Records in")), desc="Entities out")
     )
 
     mapping.dumper.close()
