+++ thebeast/tests/sample/mappings/ukrainian_mps_multithreaded.yaml
@@ -19,8 +19,7 @@ ingest:
 digest:
   cls: thebeast.digest.MultiProcessDigestor
   params:
-    processes: 1
-
+    processes: 2
   meta:
     locale:
       literal: uk
@@ -165,4 +164,4 @@ dump:
   cls: thebeast.dump.FTMLinesWriter
   params:
     # Just stdout it, so we can pipe it into the next tool
-    output_uri: "/tmp/single.jsonlines"
+    output_uri: "/tmp/multi.jsonlines"
